requirements.txt
asgiref==3.3.1
Django==3.1.4
django-cors-headers==3.5.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.4
six==1.15.0
sqlparse==0.4.1

PyJWT==1.7.1
djangorestframework-jwt==1.11.0